safari-534.53-branch/LayoutTests/fast/dom/SelectorAPI/caseID-expected.txt

PASS document.querySelector('#lower1').textContent is 'lower 1'
PASS document.querySelector('#LOWER2').textContent is 'lower 2'
PASS document.querySelector('#UPPER1').textContent is 'UPPER 1'
PASS document.querySelector('#upper2').textContent is 'UPPER 2'
PASS document.getElementById('lower1').webkitMatchesSelector('#lower1') is true
PASS document.getElementById('lower2').webkitMatchesSelector('#LOWER2') is true
PASS document.getElementById('UPPER1').webkitMatchesSelector('#UPPER1') is true
PASS document.getElementById('UPPER2').webkitMatchesSelector('#upper2') is true
PASS successfullyParsed is true

TEST COMPLETE
